cqrs app-ba
