+++ bbxBE.Application/Commands/ValidatorRules/USR_USERValidator_NEMKELL.cs
@@ -2,7 +2,6 @@
 using bbxBE.Application.Interfaces.Repositories;
 using bbxBE.Application.Wrappers;
 using bbxBE.Domain.Entities;
-using bbxBE.Infrastructure.Persistence.Contexts;
 using FluentValidation;
 using MediatR;
 using Microsoft.Extensions.Configuration;
@@ -15,7 +14,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using System.Xml.Linq;
 
-namespace bbxBE.Commands.ValidatorRules
+namespace bbxBE.Application.Commands.ValidatorRules
 {
 
     public static class Extensions
