+++ bbxBE.Application/Commands/cmdUSR_USER/DeleteUSR_USERCommand.cs
@@ -3,7 +3,6 @@ using AutoMapper.Configuration.Conventions;
 using bbxBE.Application.Interfaces.Repositories;
 using bbxBE.Application.Wrappers;
 using bbxBE.Domain.Entities;
-using bbxBE.Infrastructure.Persistence.Contexts;
 using MediatR;
 using Microsoft.Extensions.Configuration;
 using System;
@@ -13,7 +12,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace bbxBE.Commands.cmdUSR_USER
+namespace bbxBE.Application.Commands.cmdUSR_USER
 {
     public class DeleteUSR_USERCommand : IRequest<Response<long>>
     {
