+++ bbxBE.Application/Mappings/MapQueries.cs
@@ -1,15 +1,15 @@
 ﻿using AutoMapper;
+using bbxBE.Application.Queries.ViewModels;
 using bbxBE.Domain.Entities;
-using bbxBE.Queries.ViewModels;
 using System;
 using System.Collections.Generic;
 using System.Text;
 
 namespace bbxBE.Queries.Mappings
 {
-    public class GeneralProfile : Profile
+    public class MapQueries : Profile
     {
-        public GeneralProfile()
+        public MapQueries()
         {
             CreateMap<GetCustomerViewModel, Customer>().ReverseMap();
             CreateMap<GetUSR_USERViewModel, USR_USER>().ReverseMap();
