+++ bbxBE.Infrastructure.Persistence/bbxBE.Infrastructure.Persistence.csproj
@@ -20,7 +20,6 @@
   <ItemGroup>
     <ProjectReference Include="..\bbxBE.Application\bbxBE.Application.csproj" />
     <ProjectReference Include="..\bbxBE.Domain\bbxBE.Domain.csproj" />
-    <ProjectReference Include="..\bbxBE.Queires\bbxBE.Queires.csproj" />
   </ItemGroup>
 
 </Project>
