+++ bbxBE.sln
@@ -19,11 +19,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "bbxBE.Infrastructure.Persis
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "bbxBE.Infrastructure.Shared", "bbxBE.Infrastructure.Shared\bbxBE.Infrastructure.Shared.csproj", "{7E18D30F-69BC-4BF5-8640-B425DEB00D0E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "bbxBE.Commands", "bbxBE.Commands\bbxBE.Commands.csproj", "{656BBDCF-E465-4487-A2E6-AF69B2D31E6F}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "bbxBE.Queires", "bbxBE.Queires\bbxBE.Queires.csproj", "{EEA4785B-A53A-40F1-A4C9-802391157DC8}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "bbxBE.Common", "bbxBE.Common\bbxBE.Common.csproj", "{3A99ED9D-3112-41A6-9E8D-A24E7B41FE50}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "bbxBE.Common", "bbxBE.Common\bbxBE.Common.csproj", "{3A99ED9D-3112-41A6-9E8D-A24E7B41FE50}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -51,14 +47,6 @@ Global
 		{7E18D30F-69BC-4BF5-8640-B425DEB00D0E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7E18D30F-69BC-4BF5-8640-B425DEB00D0E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{7E18D30F-69BC-4BF5-8640-B425DEB00D0E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{656BBDCF-E465-4487-A2E6-AF69B2D31E6F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{656BBDCF-E465-4487-A2E6-AF69B2D31E6F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{656BBDCF-E465-4487-A2E6-AF69B2D31E6F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{656BBDCF-E465-4487-A2E6-AF69B2D31E6F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EEA4785B-A53A-40F1-A4C9-802391157DC8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EEA4785B-A53A-40F1-A4C9-802391157DC8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EEA4785B-A53A-40F1-A4C9-802391157DC8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EEA4785B-A53A-40F1-A4C9-802391157DC8}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3A99ED9D-3112-41A6-9E8D-A24E7B41FE50}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3A99ED9D-3112-41A6-9E8D-A24E7B41FE50}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3A99ED9D-3112-41A6-9E8D-A24E7B41FE50}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -73,8 +61,6 @@ Global
 		{BBBB79D2-2DE1-40E4-959E-3127BDC97916} = {D808B9C3-172D-4F1F-8F0A-10E164A3B4FC}
 		{62EFB043-0725-4876-81A9-BA2832460B4C} = {5AE26AFF-1996-4827-819C-BE113A0C830C}
 		{7E18D30F-69BC-4BF5-8640-B425DEB00D0E} = {5AE26AFF-1996-4827-819C-BE113A0C830C}
-		{656BBDCF-E465-4487-A2E6-AF69B2D31E6F} = {D808B9C3-172D-4F1F-8F0A-10E164A3B4FC}
-		{EEA4785B-A53A-40F1-A4C9-802391157DC8} = {D808B9C3-172D-4F1F-8F0A-10E164A3B4FC}
 		{3A99ED9D-3112-41A6-9E8D-A24E7B41FE50} = {D808B9C3-172D-4F1F-8F0A-10E164A3B4FC}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
